refactor: 라우터 클래스 명시적 분리 및 서버라우터 미사용 코드 제거
