+++ packages/vanilla/src/lib/index.js
@@ -1,5 +1,5 @@
 export * from "./createObserver";
 export * from "./createStorage";
 export * from "./createStore";
-export * from "./Router";
 export * from "./ServerRouter";
+export * from "./SPARouter";
